move check to uninstall + uninstall flag update check to not list by label
